refactor: add twMerge
