+++ src/svalka/Header.tsx
@@ -1,7 +1,7 @@
 import { type FC } from 'react';
 import { Link, useLocation } from 'react-router-dom';
 
-import { ThemeToggle } from './ThemeToggle.tsx';
+import { ThemeToggle } from '../features/theme/ThemeToggle.tsx';
 import { SearchBar } from '../features/search-character/ui/SearchBar.tsx';
 
 interface HeaderProps {
